Biome Assignment & Switching
As you discover a new chunk, it will be assigned a biome (e.g. texture set). These biome_number/ texture pairs are stored in a list so that they can be redrawn if you visit the biome again.

TODO: I need to make the biome transition a smooth one (i.e. One chunk changing at a time, rather than all 9 at once).
I also need to change the biome definitions so that the textures in each one are unique.
Assets such as trees, rocks, etc need to be drawn for each biome too.